# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Books API",
    "version": "1.0.0",
    "description": "A simple API for working with books"
  },
  "servers": [
    {
      "url": "/",
      "description": "Current server"
    }
  ],
  "paths": {
    "/books": {
      "get": {
        "summary": "Retrieve a list of all books",
        "description": "Fetches all book records stored in the database.",
        "tags": [
          "Books"
        ],
        "responses": {
          "200": {
            "description": "A list of books retrieved successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string",
                        "example": "b1"
                      },
                      "title": {
                        "type": "string",
                        "example": "The Great Gatsby"
                      },
                      "author": {
                        "type": "string",
                        "example": "F. Scott Fitzgerald"
                      },
                      "publicationDate": {
                        "type": "string",
                        "example": "1925-04-10"
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Internal server error"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/books/{id}": {
      "get": {
        "summary": "Retrieve a single book by ID",
        "description": "Fetches details for a specific book using its custom identifier.",
        "tags": [
          "Books"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "description": "Unique string identifier of the book (e.g., b1)",
            "schema": {
              "type": "string"
            },
            "example": "b1"
          }
        ],
        "responses": {
          "200": {
            "description": "Book details retrieved successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "example": "b1"
                    },
                    "title": {
                      "type": "string",
                      "example": "The Great Gatsby"
                    },
                    "author": {
                      "type": "string",
                      "example": "F. Scott Fitzgerald"
                    },
                    "publicationDate": {
                      "type": "string",
                      "example": "1925-04-10"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Book not found.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Book not found"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Internal server error"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/authors": {
      "get": {
        "summary": "Get all authors",
        "tags": [
          "Authors"
        ],
        "responses": {
          "200": {
            "description": "A list of authors"
          },
          "500": {
            "description": "Unable to retrieve authors"
          }
        }
      },
      "post": {
        "summary": "Create a new author",
        "tags": [
          "Authors"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "id",
                  "name",
                  "birthYear",
                  "nationality"
                ],
                "properties": {
                  "id": {
                    "type": "string",
                    "example": "a3"
                  },
                  "name": {
                    "type": "string",
                    "example": "Gabriel García Márquez"
                  },
                  "birthYear": {
                    "type": "integer",
                    "example": 1927
                  },
                  "nationality": {
                    "type": "string",
                    "example": "Colombian"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Author created successfully"
          },
          "400": {
            "description": "Invalid input or duplicate ID"
          },
          "500": {
            "description": "Unable to create author"
          }
        }
      }
    },
    "/authors/{id}": {
      "get": {
        "summary": "Get an author by custom ID",
        "tags": [
          "Authors"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "a1"
          }
        ],
        "responses": {
          "200": {
            "description": "Author found"
          },
          "404": {
            "description": "Author not found"
          },
          "500": {
            "description": "Unable to retrieve author"
          }
        }
      },
      "put": {
        "summary": "Update an existing author",
        "tags": [
          "Authors"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "a1"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "birthYear",
                  "nationality"
                ],
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "Jane Austen"
                  },
                  "birthYear": {
                    "type": "integer",
                    "example": 1775
                  },
                  "nationality": {
                    "type": "string",
                    "example": "British"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Author updated successfully"
          },
          "400": {
            "description": "Invalid input"
          },
          "404": {
            "description": "Author not found"
          },
          "500": {
            "description": "Unable to update author"
          }
        }
      },
      "delete": {
        "summary": "Delete an author",
        "tags": [
          "Authors"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "a3"
          }
        ],
        "responses": {
          "204": {
            "description": "Author deleted successfully"
          },
          "400": {
            "description": "Cannot delete author linked to existing books"
          },
          "404": {
            "description": "Author not found"
          },
          "500": {
            "description": "Unable to delete author"
          }
        }
      }
    }
  },
  "components": {},
  "tags": []
}